Break down rustpython_parser::error module
because it doesn't share any common errors but specific error for each sub module
