+++ compiler/parser/src/lib.rs
@@ -124,7 +124,6 @@
 extern crate log;
 pub use rustpython_ast as ast;
 
-pub mod error;
 mod function;
 pub mod lexer;
 pub mod mode;
