+++ src/shell.rs
@@ -1,9 +1,6 @@
 mod helper;
 
-use rustpython_parser::{
-    error::{LexicalErrorType, ParseErrorType},
-    token::Tok,
-};
+use rustpython_parser::{lexer::LexicalErrorType, parser::ParseErrorType, token::Tok};
 use rustpython_vm::{
     builtins::PyBaseExceptionRef,
     compiler::{self, CompileError, CompileErrorBody, CompileErrorType},
